requirements.txt
openai>=1.60.0
pydantic>=2.0.0
pyyaml>=6.0
phonenumbers>=8.13.0
qdrant-client>=1.12.0
fastembed>=0.4.0
rank-bm25>=0.2.2
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.0
